pyproject.toml
[tool.poetry]
name = "spotify-top-songs"
version = "0.1.0"
description = ""
authors = ["keatonconrad <keaton.conrad@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
sqlalchemy = "^2.0.36"
spotipy = "^2.24.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
tqdm = "^4.67.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
